rag_text2sql/requirements.txt
langchain-core==0.1.40
langchain-community==0.0.29
langchain-groq==0.1.1
sqlalchemy==2.0.25
streamlit==1.29.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pypdf2==3.0.1
python-dotenv==1.0.0